Allow function expression in an expression statement
Bugfix: allow function definitions to be the expression in expression statements. Prior to this commit, `fn a() {};` as well as `fn () {};` were invalid, as the parser did not consume semicolons after function definitions. This commit fixes this by checking for semicolons at the beginning of statements immediately following a function statement, converting it to an expression statement.